Conflicts:
	Team8SMS/src/main/java/sg/nus/iss/team8/demo/controllers/AdminController.java
	Team8SMS/src/main/java/sg/nus/iss/team8/demo/services/AdminService.java
	Team8SMS/src/main/java/sg/nus/iss/team8/demo/services/AdminServiceImpl.java

+++ Team8SMS/src/main/java/sg/nus/iss/team8/demo/controllers/AdminController.java
@@ -1,5 +1,6 @@
 package sg.nus.iss.team8.demo.controllers;
 
+import java.io.IOException;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -13,6 +14,8 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -43,20 +46,25 @@ import sg.nus.iss.team8.demo.models.Status;
 import sg.nus.iss.team8.demo.models.Student;
 import sg.nus.iss.team8.demo.services.AdminService;
 import sg.nus.iss.team8.demo.services.AdminServiceImpl;
-import sg.nus.iss.team8.demo.services.FacultyService;
-import sg.nus.iss.team8.demo.services.FacultyServiceImplementation;
-import sg.nus.iss.team8.demo.services.StudentService;
+import sg.nus.iss.team8.demo.services.GenerateReportService;
+import sg.nus.iss.team8.demo.services.GenerateReportServiceImpl;
 
 
 @Controller
 public class AdminController {
 
 	private AdminService aService;
+	private GenerateReportService grs;
 
 	// injection of faculty service
 	@Autowired
 	public void setAdminService(AdminServiceImpl aService) {
 		this.aService = aService;
+	}
+	
+	@Autowired
+	public void setGrs(GenerateReportServiceImpl grs) {
+		this.grs = grs;
 	}
 
 	@InitBinder
@@ -121,12 +129,27 @@ public class AdminController {
 		return "redirect:/facultymanagement";
 	}
 	
-	public static final String CURRENTSEMESTER = "AY2019/2020Sem2";
-
 	@GetMapping("/administrator")
 	public String getAdmin(Model model) {
-		model.addAttribute("currentsemester", CURRENTSEMESTER);
+		
+		ArrayList<Semester> allSemesters = aService.findAllSemsters();
+		Semester currentSemester = aService.currentSemester();
+		allSemesters.add(currentSemester);
+		
+		model.addAttribute("allSemesters", allSemesters);
+		model.addAttribute("currentSemester", currentSemester);
 		return "administrator";
+	}
+	
+	@PostMapping("/savecurrentsemester")
+	public String saveCurrentSemester(@ModelAttribute Semester sem, BindingResult bindingResult) {
+		if (bindingResult.hasErrors()) {
+			return "administrator";
+		}
+		aService.saveSemester(sem);
+		int graduationThreshold = 20;
+		aService.applyGraduatedStatus(sem, graduationThreshold);
+		return "redirect:/administrator";
 	}
 
 	@GetMapping("/studentmanagement")
@@ -203,7 +226,7 @@ public class AdminController {
 	}
 	
 	@GetMapping("/{acceptOrReject}/{id}/{courseCode}/{semesterid}")
-	public String rejectCourseApplication(@PathVariable(name="acceptOrReject") String acceptOrReject,
+	public String acceptOrRejectCourseApplication(@PathVariable(name="acceptOrReject") String acceptOrReject,
 			@PathVariable(name="id") int id, 
 			@PathVariable(name="courseCode") String courseCode,
 			@PathVariable(name="semesterid") int semesterid) {
@@ -250,11 +273,17 @@ public class AdminController {
 	}
 
 	@PostMapping("/savecourserun")
-	public String saveCourserun(@Valid @ModelAttribute Courserun course, BindingResult bindingResult) {
+	public String saveCourserun(@Valid @ModelAttribute Courserun course, @RequestParam("currentSemester")String semLabel, BindingResult bindingResult) {
 		if (bindingResult.hasErrors()) {
 			return "courserunform";
 		}
-
+		if(semLabel != "") {
+			aService.findOrAddSemester(semLabel);
+			course.setSemester(aService.findOrAddSemester(semLabel));
+			String shortSemLabel = concatSemester(semLabel);
+			course = aService.concatCourseNameWithYear(course, shortSemLabel);
+		}
+		
 		aService.saveCourserun(course);
 		return "redirect:/courserunmanagement";
 	}
@@ -262,16 +291,12 @@ public class AdminController {
 	@GetMapping("/addcourserun")
 	public String showAddCourseForm(Model model) {
 		Courserun course = new Courserun();
-		Semester currentSemester = aService.currentSemester();
+		String semLabel=null;
 		ArrayList<Faculty> flist = aService.findAllFaculty();
 		model.addAttribute("flist", flist);
 		model.addAttribute("course", course);
-		model.addAttribute("currentSemester", currentSemester);
-
-		String longSemLabel = currentSemester.getLabel();
-		String shortSemLabel = concatSemester(longSemLabel);
-		model.addAttribute("shortSemLabel", shortSemLabel);
-
+		model.addAttribute("currentSemester", semLabel);
+		
 		return "courserunform";
 	}
 
@@ -289,8 +314,15 @@ public class AdminController {
 	public String deleteCourserun(Model model, @PathVariable(name = "courseCode") String courseCode,
 			@PathVariable(name = "semesterid") int semesterid) {
 		Courserun course = aService.findCourserun(courseCode, semesterid);
-		aService.removeCourserun(course);
-		return "redirect:/courserunmanagement";
+		// if courserunstudents is not empty, redirect to /viewcourserun/"+courseCode+"/"+semesterid;
+		ArrayList<CourserunStudent> studentsPerCourse = aService.findStudentsByCourseName(course.getCourseName());
+		if(studentsPerCourse.isEmpty()) {
+			aService.removeCourserun(course);
+			return "redirect:/courserunmanagement";
+		} else {
+			return "redirect:/viewcourserun/"+courseCode+"/"+semesterid;
+		}
+		
 	}
 
 	@GetMapping("/viewcourserun/{courseCode}/{semesterid}")
@@ -387,6 +419,34 @@ public class AdminController {
 		model.addAttribute("yearMonths", yearMonths);
 		model.addAttribute("selectedmonth",ym);
 		return "admin_movementregister";
+	}
+	
+	@RequestMapping("/downloadCSV/classlist")
+	public void downloadCSV(HttpServletRequest request, HttpServletResponse response, 
+			@RequestParam("coursename") String coursename) throws IOException {
+		// create a list of object
+		
+			System.out.println("found " + coursename);
+			ArrayList<CourserunStudent> students = aService.findStudentsByCourseName(coursename);
+			System.out.println("found: " + students.size() + " students");
+			System.out.println("trying to download");
+			String[] headers=new String[]{"Course Name","Student Id", "Student Name", "Degree", "Mobile", "Email", "Grade", "Status"};
+			System.out.println("request: " + request);
+			System.out.println("response: " + response);
+			
+			grs.ExportCSV(request, response, students, headers);
+	}
+	
+	@GetMapping("/viewstudentcourses/{studentid}")
+	public String viewCourserun(Model model,
+			@PathVariable(name = "studentid") int studentid) {
+		Student student = aService.findStudent(studentid);
+		ArrayList<CourserunStudent> coursesPerStudent = aService.findCoursesByStudentId(studentid);
+
+		model.addAttribute("student", student);
+		model.addAttribute("coursesPerStudent", coursesPerStudent);
+
+		return "viewstudentcourses";
 	}
 
 }
